Removed stl lists from parse_bmp

+++ include/parse_utils.h
@@ -18,6 +18,6 @@ enum parse_msg_error {
     //TODO : More error types to come
 };
 
-ssize_t extract_from_buffer (unsigned char*& buffer, int &buf_len, void *output_buf, int output_len);
+ssize_t extract_from_buffer (unsigned char*& buffer, int &buf_len, void *output_buf, ssize_t output_len);
 
 #endif //PARSE_LIB_PARSEUTILS_H
